crowd-edge-cloud/services/dashboard/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-cloud-firestore==2.15.0
